+++ Sources/Flash/Extensions/LeafTagConfig.swift
@@ -1,0 +1,7 @@
+import Leaf
+
+public extension LeafTagConfig {
+    public mutating func useFlashLeafTags() {
+        use(FlashTag(), as: "flash")
+    }
+}
